POC how to avoid no time limit on getting next graph packet
Idea is to check for the packet queue size in output stream and only
call Next when we know for sure that there is packet.

It doesn't work for positive path (QueueSize doesn't show that there is
packet waiting to be extracted). On negative path we most probably would
wait for full timeout to return error code.